+++ src/qbrsettingsdialog.cpp
@@ -3,6 +3,9 @@
 #include "qbrcfg.h"
 
 #include <QColorDialog>
+#include <QFileDialog>
+#include <QUrl>
+#include <QDebug>
 
 QBRSettingsDialog::QBRSettingsDialog(QWidget *parent, Qt::WindowFlags f) :
     QDialog(parent, f),
@@ -38,11 +41,19 @@ void QBRSettingsDialog::settingsLoad()
 
     bookBgColor = qbrcfg::getBookBgColor();
     findChild<QPushButton*>("bgColorButton")->setStyleSheet(QString("background-color: %1").arg(bookBgColor));
+
+    findChild<QCheckBox*>("customCssEnable")->setChecked(qbrcfg::getCustomStyleEnabled());
+    findChild<QLineEdit*>("customCssUrl")->setEnabled(qbrcfg::getCustomStyleEnabled());
+    findChild<QLineEdit*>("customCssUrl")->setText(qbrcfg::getCustomStyleUrl());
+    findChild<QPushButton*>("customCssSelect")->setEnabled(qbrcfg::getCustomStyleEnabled());
 }
 void QBRSettingsDialog::settingsSave()
 {
     qbrcfg::setStatusBarEnabled(findChild<QCheckBox*>("statusBarEnable")->isChecked());
     qbrcfg::setLastOpenedFileEnable(findChild<QCheckBox*>("enableLoadLast")->isChecked());
+
+    qbrcfg::setCustomStaticEnabled(findChild<QCheckBox*>("customCssEnable")->isChecked());
+    qbrcfg::setCustomStyleUrl(findChild<QLineEdit*>("customCssUrl")->text());
 
     if (findChild<QRadioButton*>("uiMenuOnly")->isChecked())
     {
@@ -66,6 +77,35 @@ void QBRSettingsDialog::bgColorChoose()
         bookBgColor = color.name();
         findChild<QPushButton*>("bgColorButton")->setStyleSheet(QString("background-color: %1").arg(bookBgColor));
     }
+}
+
+void QBRSettingsDialog::customCssSelect()
+{
+    QString fileName = QFileDialog::getOpenFileName(
+        this,
+        tr("Select File"),
+        QDir::homePath(),
+        tr("Stylesheets (%1)").arg("*.css")
+    );
+    if (fileName != "")
+    {
+        findChild<QLineEdit*>("customCssUrl")->setText(QUrl::fromLocalFile(fileName).toString());
+    }
+}
+
+void QBRSettingsDialog::customCssEnable(int state)
+{
+    if (state == Qt::Checked)
+    {
+        findChild<QLineEdit*>("customCssUrl")->setEnabled(true);
+        findChild<QPushButton*>("customCssSelect")->setEnabled(true);
+    }
+    else if (state == Qt::Unchecked)
+    {
+        findChild<QLineEdit*>("customCssUrl")->setEnabled(false);
+        findChild<QPushButton*>("customCssSelect")->setEnabled(false);
+    }
+
 }
 
 QBRSettingsDialog::~QBRSettingsDialog()